added update component
