+++ client/src/App.js
@@ -4,6 +4,7 @@ import Login from './screens/Login/Login'
 import WelcomePage from './screens/WelcomePage/WelcomePage';
 import Services from './screens/Services/Services'
 import ShowAppointment from './screens/ShowAppointment/ShowAppointment'
+import UpdateDeleteAppointment from './screens/UpdateDeleteAppointment/UpdateDeleteAppointment'
 import CreateAppointment from './screens/CreateAppointment/CreateAppointment';
 import Calendar from './components/Calendar/Calendar'
 
@@ -14,8 +15,9 @@ export default function App(props) {
       {/* <WelcomePage/> */}
       {/* <Services /> */}
       {/* <ShowAppointment /> */}
-      <CreateAppointment/>
+      {/* <CreateAppointment/> */}
       {/* <Calendar/> */}
+      <UpdateDeleteAppointment/>
       </div>
   );
 }
